BigDecimal in POS (complete). Integration of ofbiz-1945.patch changes (with some modifications) but still not used since it miss some features: * AutoSave (automatically save current transaction) * Add a saved sale to the current sale * Remove a savec sale * Replace current sale by a saved sale and remove simultaneously the saved sale
git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@745257 13f79535-47bb-0310-9956-ffa450edef68

+++ specialpurpose/pos/src/org/ofbiz/pos/component/Operator.java
@@ -103,7 +103,7 @@ public class Operator {
         if (OPER_TOTAL[0].equals(fieldName)) {
             String total = "0.00";
             if (trans != null) {
-                total = UtilFormatOut.formatPrice(trans.getTotalDue().doubleValue());
+                total = UtilFormatOut.formatPrice(trans.getTotalDue());
             }
             field.setText(total);
         } else if (OPER_DATE[0].equals(fieldName)) {
